Reformat + Update
Reformat the robot_plugin.cpp through CLion
Update robot_plugin to fix the issue of having 0 torque set on iterations where torque is not set. This is not obvious when reading from the JointPtr->GetForce() as it will return a previously set force. However, even when it returns a positive value for the GetForce(), there is 0 angular acceleration which means 0 torque. This means that JointPtr->GetForce() is misleading.